+++ ice/src/util/mod.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod util_test;
 
-use crate::agent::agent_config::InterfaceFilterFn;
+use crate::agent::agent_config::{InterfaceFilterFn, IpFilterFn};
 use crate::error::*;
 use crate::network_type::*;
 
@@ -90,6 +90,7 @@ pub async fn stun_request(
 pub async fn local_interfaces(
     vnet: &Arc<Net>,
     interface_filter: &Option<InterfaceFilterFn>,
+    ip_filter: &Option<IpFilterFn>,
     network_types: &[NetworkType],
 ) -> HashSet<IpAddr> {
     let mut ips = HashSet::new();
@@ -114,11 +115,20 @@ pub async fn local_interfaces(
 
         for ipnet in iface.addrs() {
             let ipaddr = ipnet.addr();
+
             if !ipaddr.is_loopback()
                 && ((ipv4requested && ipaddr.is_ipv4()) || (ipv6requested && ipaddr.is_ipv6()))
             {
-                ips.insert(ipaddr);
+                continue;
             }
+
+            if let Some(filter) = ip_filter {
+                if !filter(ipaddr) {
+                    continue;
+                }
+            }
+
+            ips.insert(ipaddr);
         }
     }