+++ ice/CHANGELOG.md
@@ -1,8 +1,13 @@
 # webrtc-ice changelog
 
 ## Unreleased
+* Add IP filter to ICE `AgentConfig` [#306](https://github.com/webrtc-rs/webrtc/pull/306)
 
-### v0.8.0
+## v0.8.1
+
+This release was released in error and contains no changes from 0.8.0.
+
+## v0.8.0
 
 * Increased min verison of `log` dependency to `0.4.16`. [#250 Fix log at ^0.4.16 to make tests compile](https://github.com/webrtc-rs/webrtc/pull/250) by [@k0nserv](https://github.com/k0nserv).
 * Incresed serde's minimum version to 1.0.102 [#243 Fixes for cargo minimal-versions](https://github.com/webrtc-rs/webrtc/pull/243) contributed by [algesten](https://github.com/algesten)